文件更新

+++ lab5/libs/defs.h
@@ -10,7 +10,7 @@
 #define __noreturn __attribute__((noreturn))
 
 /* Represents true-or-false values */
-typedef long long bool;
+typedef int bool;
 
 /* Explicitly-sized versions of integer types */
 typedef char int8_t;
